+++ hypervisor/Makefile
@@ -99,9 +99,7 @@ INCLUDE_PATH += include/arch/x86
 INCLUDE_PATH += include/arch/x86/guest
 INCLUDE_PATH += include/debug
 INCLUDE_PATH += include/public
-ifeq ($(CONFIG_PARTITION_MODE),y)
 INCLUDE_PATH += include/dm
-endif
 INCLUDE_PATH += bsp/include
 INCLUDE_PATH += bsp/include/$(CONFIG_PLATFORM)
 INCLUDE_PATH += boot/include
